Refactor routes with updated index
