+++ db.js
@@ -1,0 +1,6 @@
+const sqlite3 = require('sqlite3').verbose();
+const path = require('path');
+
+const db = new sqlite3.Database(path.join(__dirname, 'raw_articles.db'));
+
+module.exports = db;
