+++ lib/extractParties.js
@@ -62,4 +62,21 @@ function getFirstSentence(text) {
   return sentence.trim();
 }
 
-module.exports = { parseOpenAIResponse, getFirstSentence };
+async function extractParties(openai, title, body) {
+  const firstSentence = getFirstSentence(body);
+  const titleAndSentence = `${title || ''} ${firstSentence}`.trim();
+  const prompt =
+    `Extract the acquiror and target from this text. If none are mentioned, respond with {"acquiror":"N/A","target":"N/A"}. Text: "${titleAndSentence}"`;
+
+  const resp = await openai.chat.completions.create({
+    model: 'gpt-3.5-turbo',
+    messages: [{ role: 'user', content: prompt }],
+    temperature: 0
+  });
+
+  const output = resp.choices[0].message.content.trim();
+  const { acquiror, target } = parseOpenAIResponse(output);
+  return { acquiror, target, prompt, firstSentence, output };
+}
+
+module.exports = { parseOpenAIResponse, getFirstSentence, extractParties };
