+++ routes/articles.js
@@ -1,0 +1,88 @@
+const express = require('express');
+const { OpenAI } = require('openai');
+const db = require('../db');
+const fetchBody = require('../lib/enrichment/fetchBody');
+const extractParties = require('../lib/enrichment/extractParties');
+
+const router = express.Router();
+const openai = new OpenAI({ apiKey: process.env.OPENAI_API_KEY });
+
+// Get all articles
+router.get('/', (req, res) => {
+  const query = `
+    SELECT
+      a.id,
+      a.title,
+      a.description,
+      a.time,
+      a.link,
+      GROUP_CONCAT(f.id) as filter_ids,
+      GROUP_CONCAT(f.name) as filter_names
+    FROM articles a
+    LEFT JOIN article_filter_matches m ON a.id = m.article_id
+    LEFT JOIN filters f ON f.id = m.filter_id
+    GROUP BY a.id
+    ORDER BY a.created_at DESC`;
+
+  db.all(query, [], (err, rows) => {
+    if (err) {
+      console.error(err);
+      return res.status(500).json({ error: 'Failed to retrieve articles' });
+    }
+    rows.forEach(r => {
+      r.filter_ids = r.filter_ids
+        ? r.filter_ids.split(',').map(id => parseInt(id, 10))
+        : [];
+      r.filter_names = r.filter_names ? r.filter_names.split(',') : [];
+      r.matched = r.filter_ids.length > 0;
+    });
+    res.json(rows);
+  });
+});
+
+// Get today's articles that matched the M&A filter
+router.get('/mna-today', (req, res) => {
+  const query = `
+    SELECT a.id, a.title, a.description, a.time, a.link,
+           ae.body, ae.acquiror, ae.target
+    FROM articles a
+    JOIN article_filter_matches m ON a.id = m.article_id
+    JOIN filters f ON f.id = m.filter_id
+    LEFT JOIN article_enrichments ae ON a.id = ae.article_id
+    WHERE date(a.created_at) = date('now') AND f.name = 'M&A'
+    ORDER BY a.created_at DESC`;
+
+  db.all(query, [], (err, rows) => {
+    if (err) {
+      console.error(err);
+      return res.status(500).json({ error: 'Failed to retrieve articles' });
+    }
+    res.json(rows);
+  });
+});
+
+// Enrich an article by scraping its body text
+router.post('/:id/enrich', async (req, res) => {
+  const { id } = req.params;
+  try {
+    const text = await fetchBody(db, id);
+    res.json({ success: true, body: text });
+  } catch (err) {
+    console.error(err);
+    res.status(500).json({ error: 'Failed to enrich article' });
+  }
+});
+
+// Extract acquiror and target using GPT-3.5 from the first sentence
+router.post('/:id/extract-parties', async (req, res) => {
+  const { id } = req.params;
+  try {
+    const result = await extractParties(db, openai, id);
+    res.json({ success: true, ...result });
+  } catch (err) {
+    console.error(err);
+    res.status(500).json({ error: 'Failed to extract parties' });
+  }
+});
+
+module.exports = router;
